backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
jwt==1.3.1
scikit-learn==1.3.2
numpy==1.26.2
pandas==2.1.3
tensorflow==2.15.0
torch==2.1.1
transformers==4.35.2
langchain==0.1.0
chromadb==0.4.21
sentence-transformers==2.2.2
python-dotenv==1.0.0
cors==1.0.1
cors-starlette==0.0.2
slowapi==0.1.9
logging-loki==0.3.2
